requirements.txt
aiogram==3.4.1
requests==2.31.0
beautifulsoup4==4.12.3
APScheduler==3.10.4
matplotlib==3.8.4
python-dotenv==1.0.1